Remove IUpdateSqlGenerator from Migrations
